backend/app/db.py
from pathlib import Path
from collections.abc import Generator
from sqlalchemy import event
from sqlmodel import SQLModel, Session, create_engine

BASE_DIR = Path(__file__).resolve().parents[2]  # .../inetum
DB_PATH = BASE_DIR / "app.db"
DATABASE_URL = f"sqlite:///{DB_PATH}"

engine = create_engine(
    DATABASE_URL, echo=False, connect_args={"check_same_thread": False}
)


@event.listens_for(engine, "connect")
def _set_sqlite_pragma(dbapi_connection, connection_record) -> None:
    # Asegura integridad referencial en SQLite
    cursor = dbapi_connection.cursor()
    cursor.execute("PRAGMA foreign_keys=ON")
    cursor.close()


def init_db() -> None:
    """Create all tables."""
    from . import models  # noqa: F401

    SQLModel.metadata.create_all(bind=engine)


def get_session() -> Generator[Session, None, None]:
    with Session(engine) as session:
        yield session

