requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.0
python-multipart==0.0.9
email-validator>=2.0.0
openpyxl==3.1.5
